Move components to folder 'components' in 'src' from 'src'
프로젝트를 확장할때, 다른 속성의 파일들이 많아질 경우를 고려하여 컴퍼넌트들을 'components'파일로 따로 분류하였습니다.
